Start Comntributer image with build
